+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This release changes the install directory from ~/Mycodo to /opt/Mycodo. This ne
 ### Features
 
  - Add Output: GPIO On/Off using pinctrl (First Pi 5-compatible Output)
+ - Add Output: PWM MQTT Publish
  - Add ability to switch displaying hostname with custom text
  - Add Step Line Series Type to Graph (Synchronous) Widget
  - Add controller_restart as client endpoint